refactor: move other utils to utils file
